requirements.txt
# ─── Core LLM ───────────────────────────────────────────
anthropic==0.25.1
openai==1.30.1                  # fallback / embeddings
google-generativeai==0.7.2
# ─── Agent Framework ────────────────────────────────────
langchain
langchain-anthropic
langchain-community
langchain-core

# ─── RAG / Vector Store ─────────────────────────────────
faiss-cpu==1.8.0
sentence-transformers==3.0.0
tiktoken==0.7.0

# ─── Document Handling ──────────────────────────────────
pypdf==4.2.0
python-docx==1.1.2
markdown==3.6
reportlab==4.2.0               # PDF generation
jinja2==3.1.4                  # Document templating

# ─── Data & Utils ───────────────────────────────────────
python-dotenv==1.0.1
pydantic==2.7.1
pandas==2.2.2
numpy==1.26.4
rich==13.7.1                   # Pretty terminal output
loguru==0.7.2                  # Logging
tenacity==8.3.0                # Retry logic

# ─── API & Web ──────────────────────────────────────────
httpx==0.27.0
fastapi==0.111.0               # Optional API layer
uvicorn==0.30.0

# ─── Testing ────────────────────────────────────────────
pytest==8.2.0
pytest-asyncio==0.23.7